1. update contact ui to enable recall single sent request, accept/refuse single pending request, and a button hold for the chat page 2. enable the contact ui for the care receiver, tested contact request recall/accept/refuse from both care receiver and care giver side
