+++ app/build.gradle
@@ -126,6 +126,9 @@ dependencies {
     implementation("com.firebaseui:firebase-ui-auth:8.0.2")
     implementation("com.firebaseui:firebase-ui-database:8.0.2")
 
+    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:1.5.0"
+    implementation "com.google.firebase:firebase-firestore-ktx:23.0.0"
+
     implementation 'androidx.work:work-runtime-ktx:2.7.0'
 
     implementation("io.ktor:ktor-client-okhttp:2.3.2")
