+++ app/src/main/java/com/ece452/pillmaster/PillMasterNavHost.kt
@@ -6,17 +6,17 @@ import androidx.annotation.RequiresApi
 import com.ece452.pillmaster.screen.carereceiver.AutoFillEntry
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavGraph.Companion.findStartDestination
 import androidx.navigation.NavHostController
 import androidx.navigation.NavType
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.navArgument
-import com.ece452.pillmaster.screen.caregiver.CareGiverMessageScreen
 import com.ece452.pillmaster.screen.carereceiver.CalendarScreen
 import com.ece452.pillmaster.screen.carereceiver.CaregiverManageScreen
 import com.ece452.pillmaster.screen.carereceiver.ChatScreen
-import com.ece452.pillmaster.screen.carereceiver.MessageScreen
+import com.ece452.pillmaster.screen.common.ContactScreen
 import com.ece452.pillmaster.screen.carereceiver.PillManageScreen
 import com.ece452.pillmaster.screen.carereceiver.ReceiverSettingScreen
 import com.ece452.pillmaster.screen.common.CareGiverHomeScreen
@@ -27,6 +27,8 @@ import com.ece452.pillmaster.screen.common.LoginScreen
 import com.ece452.pillmaster.screen.common.PillAddPageScreen
 import com.ece452.pillmaster.screen.common.SignupScreen
 import com.ece452.pillmaster.utils.NavigationPath
+import com.ece452.pillmaster.viewmodel.CareGiverContactViewModel
+import com.ece452.pillmaster.viewmodel.CareReceiverContactViewModel
 
 @SuppressLint("NewApi")
 @RequiresApi(Build.VERSION_CODES.O)
@@ -71,9 +73,10 @@ fun PillMasterNavHost(
                 navController = navController
             )
         }
-        composable(NavigationPath.CARE_GIVER_MESSAGE.route) {
-            CareGiverMessageScreen(
-                navController = navController
+        composable(NavigationPath.CARE_GIVER_CONTACT.route) {
+            ContactScreen(
+                navController = navController,
+                messageViewModel = hiltViewModel<CareGiverContactViewModel>()
             )
         }
         composable(NavigationPath.PILL_ADD_PAGE.route) { entry ->
@@ -88,9 +91,10 @@ fun PillMasterNavHost(
                 navController = navController,
             )
         }
-        composable(NavigationPath.MESSAGE.route) {
-            MessageScreen(
+        composable(NavigationPath.CARE_RECEIVER_CONTACT.route) {
+            ContactScreen(
                 navController = navController,
+                messageViewModel = hiltViewModel<CareReceiverContactViewModel>()
             )
         }
         composable(NavigationPath.RECEIVER_SETTING.route) {
