+++ app/src/main/java/com/ece452/pillmaster/di/AppModule.kt
@@ -74,14 +74,13 @@ class AppModule {
 
         @Binds
         @Singleton
-        @Named("careReceiver")
+        @Named("careReceiverContact")
         fun provideCareReceiverContactRepository(repo: CareReceiverContactRepository) : IContactRepository
 
         @Binds
         @Singleton
-        @Named("careGiver")
+        @Named("careGiverContact")
         fun provideCareGiverContactRepository(repo: CareGiverContactRepository) : IContactRepository
-
 
 
         @Binds
