+++ app/src/main/java/com/ece452/pillmaster/screen/caregiver/CareGiverHomeScreen.kt
@@ -1,35 +1,25 @@
 package com.ece452.pillmaster.screen.common
 
 import android.os.Build
-import android.widget.Toast
 import androidx.annotation.RequiresApi
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Add
 import androidx.compose.material.icons.rounded.Email
 import androidx.compose.material.icons.rounded.Settings
-import androidx.compose.material3.AlertDialog
-import androidx.compose.material3.Button
 import androidx.compose.material3.Card
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextField
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -42,16 +32,12 @@ import androidx.compose.ui.semantics.contentDescription
 import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.NavController
 import coil.annotation.ExperimentalCoilApi
 import coil.compose.rememberImagePainter
 import com.ece452.pillmaster.R
 import com.ece452.pillmaster.model.Contact
 import com.ece452.pillmaster.utils.NavigationPath
-import com.ece452.pillmaster.viewmodel.CareGiverMessageViewModel
-import com.ece452.pillmaster.viewmodel.CareReceiverMessageViewModel
 
 // Sample data of receivers
 val receiverList = listOf(
@@ -105,18 +91,19 @@ val receiverList = listOf(
     "CareReceiver L"
 )
 
+
 @OptIn(ExperimentalMaterial3Api::class)
 @RequiresApi(Build.VERSION_CODES.O)
 @Composable
 fun CareGiverHomeScreen(
     // TODO - Expose an action if this action takes the user to another screen.
     navController: NavController,
-    messageViewModel: CareGiverMessageViewModel = hiltViewModel()
+    /*messageViewModel: CareGiverMessageViewModel = hiltViewModel()*/
 ) {
 
-    val connectedContacts = messageViewModel.connectedContacts.collectAsStateWithLifecycle(emptyList())
+/*    val connectedContacts = messageViewModel.connectedContacts.collectAsStateWithLifecycle(emptyList())
     val messageUiState = messageViewModel.messageUiState
-    val isError = messageUiState.error != null
+    val isError = messageUiState.error != null*/
     val context = LocalContext.current
     var showDialog by remember { mutableStateOf(false) }
 
@@ -128,7 +115,8 @@ fun CareGiverHomeScreen(
         verticalArrangement = Arrangement.Center,
         horizontalAlignment = Alignment.CenterHorizontally,
     ) {
-        Text(text = "Connected Contacts", fontSize = 24.sp, modifier = Modifier.padding(start = 10.dp))
+        ShortNavBar(navController)
+        /*Text(text = "Connected Contacts", fontSize = 24.sp, modifier = Modifier.padding(start = 10.dp))
         LazyColumn(modifier = Modifier
             .weight(1f)
             .padding(16.dp)) {
@@ -180,7 +168,7 @@ fun CareGiverHomeScreen(
         if (isError) {
             Toast.makeText(context, messageUiState.error ?: "unknown error", Toast.LENGTH_SHORT).show()
             messageUiState.error = null
-        }
+        }*/
     }
 }
 
@@ -228,8 +216,8 @@ fun ShortNavBar(
             .padding(0.dp, 0.dp),
         verticalAlignment = Alignment.CenterVertically
     ) {
-        NavItem(Icons.Rounded.Email, "Message", Color(0xFF227EBA)) {
-            navController.navigate(NavigationPath.CARE_GIVER_MESSAGE.route)
+        NavItem(Icons.Rounded.Email, "Contact", Color(0xFF227EBA)) {
+            navController.navigate(NavigationPath.CARE_GIVER_CONTACT.route)
         }
         NavItem(Icons.Rounded.Settings,"Setting", Color(0xFF227EBA) ) {
             // TODO go to care giver's manage page + sign out.
