+++ app/src/main/java/com/ece452/pillmaster/screen/carereceiver/CareReceiverHomepageScreen.kt
@@ -119,8 +119,8 @@ fun NavBar(
         NavItem(Icons.Rounded.DateRange, "Calender", Color(0xFF227EBA)) {
             navController.navigate(NavigationPath.CALENDAR.route)
         }
-        NavItem(Icons.Rounded.Email, "Message", Color(0xFF227EBA)) {
-            navController.navigate(NavigationPath.MESSAGE.route)
+        NavItem(Icons.Rounded.Email, "Contact", Color(0xFF227EBA)) {
+            navController.navigate(NavigationPath.CARE_RECEIVER_CONTACT.route)
         }
         NavItem(Icons.Rounded.Face, "ChatBot", Color(0xFF227EBA)) {
             navController.navigate(NavigationPath.HEALTH_BOT_PATH.route)
