+++ app/src/main/java/com/ece452/pillmaster/utils/NavigationPath.kt
@@ -10,7 +10,8 @@ enum class NavigationPath(val route: String) {
     PILL_ADD_PAGE("pill_add_page"),
     CAMERA_HOMEPAGE("camera_homepage"),
     CALENDAR("calendar"),
-    MESSAGE("message"),
+    CARE_GIVER_CONTACT("care_giver_contact"),
+    CARE_RECEIVER_CONTACT("care_receiver_contact"),
     RECEIVER_SETTING("receiver_setting"),
     PILL_MANAGE("pill_manage"),
     CAREGIVER_MANAGE("caregiver_manage"),
